requirements.txt
torch==1.12.1
transformers==4.23.1
piper-tts==1.0.0
langchain==0.0.161
langchain_community==0.2.0
python-dotenv==1.0.0
requests==2.31.0
fuzzywuzzy==0.18.0
